fm - Move Message class to network package.
